+++ ros/src/des_velocity_pub/package.xml
@@ -1,13 +1,13 @@
 <?xml version="1.0"?>
 <package format="2">
-  <name>speed_designer</name>
+  <name>des_velocity_pub</name>
   <version>0.0.0</version>
-  <description>The speed_designer package</description>
+  <description>The des_velocity_pub package</description>
 
   <!-- One maintainer tag required, multiple allowed, one person per tag -->
   <!-- Example:  -->
   <!-- <maintainer email="jane.doe@example.com">Jane Doe</maintainer> -->
-  <maintainer email="peng@todo.todo">peng</maintainer>
+  <maintainer email="robotpengxu@gmail.com">peng</maintainer>
 
 
   <!-- One license tag required, multiple allowed, one license per tag -->
